requirements.txt
# Core requirements
openai>=1.0.0
pydantic>=2.0.0
rich>=13.0.0
pydantic-ai>=0.0.31

# Optional AI providers
google-generativeai>=0.1.0  # For Gemini
anthropic>=0.15.0  # For Claude
ollama>=0.1.0  # For Ollama

# Other dependencies
requests>=2.31.0
python-dotenv>=1.0.0

# Core dependencies
pybind11>=2.10.0
numpy>=1.20.0
pandas>=1.3.0
rich>=10.0.0

# Visualization libraries
matplotlib>=3.5.0
seaborn>=0.11.0

# Database-related
duckdb>=0.7.0

# Utility libraries
python-dateutil>=2.8.2
tqdm>=4.62.0 